Rename Aliases to Common
