+++ src/Algos.cc
@@ -179,7 +179,7 @@ findSimilarStr(
 
 #ifdef POAC_TEST
 
-#  include "Rustify/Aliases.hpp"
+#  include "Rustify/Common.hpp"
 #  include "Rustify/Tests.hpp"
 
 #  include <array>
