+++ src/Manifest.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Rustify/Aliases.hpp"
+#include "Rustify/Common.hpp"
 #include "Semver.hpp"
 
 #include <compare>
