+++ src/Rustify.hpp
@@ -1,5 +1,5 @@
 #pragma once
 
-#include "Rustify/Aliases.hpp"
+#include "Rustify/Common.hpp"
 #include "Rustify/Tests.hpp"
 #include "Rustify/Traits.hpp"
