+++ src/VersionReq.cc
@@ -1,7 +1,7 @@
 #include "VersionReq.hpp"
 
 #include "Exception.hpp"
-#include "Rustify/Aliases.hpp"
+#include "Rustify/Common.hpp"
 
 #include <cctype>
 #include <cstddef>
